+++ src/app/evm_for_ccf.cpp
@@ -2,7 +2,6 @@
 // Licensed under the MIT License.
 
 // EVM-for-CCF
-#include "../queue/workerqueue.hpp"
 #include "account_proxy.h"
 #include "ds/logger.h"
 #include "ethereum_state.h"
@@ -12,7 +11,7 @@
 #include "tables.h"
 #include "tls/key_pair.h"
 #include "utils.h"
-#include "../queue/transaction.h"
+#include "../transaction/generator.h"
 // CCF
 #include "ds/hash.h"
 #include "enclave/app_interface.h"
@@ -48,7 +47,7 @@ namespace evm4ccf
     tables::Storage storage;
     tables::Results tx_results;
     TransactionTables txTables;
-    WorkerQueue workerQueue;
+
     EthereumState make_state(kv::Tx& tx)
     {
       return EthereumState(accounts.get_views(tx), tx.get_view(storage));
@@ -289,23 +288,23 @@ namespace evm4ccf
         return;
       };
 
-      auto send_privacy_policy = [this](ccf::EndpointContext& args) {
-        const auto body_j =
-          nlohmann::json::parse(args.rpc_ctx->get_request_body());
-        auto sppp = body_j.get<rpcparams::SendPrivacyPolicy>();
-        printf( "privacy_policy tx: %p \n", (void*)&args.tx);
-        // TODO: check target contract
-        PrivacyPolicyTransaction ppt(sppp);
-        auto hash = workerQueue.addModule(ppt);
+      // auto send_privacy_policy = [this](ccf::EndpointContext& args) {
+      //   const auto body_j =
+      //     nlohmann::json::parse(args.rpc_ctx->get_request_body());
+      //   auto sppp = body_j.get<rpcparams::SendPrivacyPolicy>();
+      //   printf( "privacy_policy tx: %p \n", (void*)&args.tx);
+      //   // TODO: check target contract
+      //   PrivacyPolicyTransaction ppt(sppp);
+      //   auto hash = workerQueue.addModule(ppt);
 
-        // Return success HTTP response with the result json
-        args.rpc_ctx->set_response_status(HTTP_STATUS_OK);
-        args.rpc_ctx->set_response_header(
-          http::headers::CONTENT_TYPE, http::headervalues::contenttype::TEXT);
-        args.rpc_ctx->set_response_body(
-          jsonrpc::result_response(0, eevm::to_hex_string(hash)).dump());
-        return;
-      };
+      //   // Return success HTTP response with the result json
+      //   args.rpc_ctx->set_response_status(HTTP_STATUS_OK);
+      //   args.rpc_ctx->set_response_header(
+      //     http::headers::CONTENT_TYPE, http::headervalues::contenttype::TEXT);
+      //   args.rpc_ctx->set_response_body(
+      //     jsonrpc::result_response(0, eevm::to_hex_string(hash)).dump());
+      //   return;
+      // };
 
       auto send_raw_privacy_policy_transaction = [this](ccf::EndpointContext& args) {
         const auto body_j = 
@@ -343,66 +342,66 @@ namespace evm4ccf
         return;
       };
 
-      auto send_multiPartyTransaction = [this](ccf::EndpointContext& args) {
-        CLOAK_DEBUG_FMT("request body:{}", args.rpc_ctx->get_request_body());
-        printf( "send_multiPartyTransaction tx: %p \n", (void*)&args.tx);
-        const auto body_j =
-          nlohmann::json::parse(args.rpc_ctx->get_request_body());
-        auto smp = body_j.get<rpcparams::SendMultiPartyTransaction>();
+      // auto send_multiPartyTransaction = [this](ccf::EndpointContext& args) {
+      //   CLOAK_DEBUG_FMT("request body:{}", args.rpc_ctx->get_request_body());
+      //   printf( "send_multiPartyTransaction tx: %p \n", (void*)&args.tx);
+      //   const auto body_j =
+      //     nlohmann::json::parse(args.rpc_ctx->get_request_body());
+      //   auto smp = body_j.get<rpcparams::SendMultiPartyTransaction>();
         
 
-        MultiPartyTransaction mpt(smp);
-        mpt.checkSignature();
-        auto result = workerQueue.addMultiParty(mpt);
+      //   MultiPartyTransaction1 mpt(smp);
+      //   mpt.checkSignature();
+      //   auto result = workerQueue.addMultiParty(mpt);
 
-        args.rpc_ctx->set_response_status(HTTP_STATUS_OK);
-        args.rpc_ctx->set_response_header(
-          http::headers::CONTENT_TYPE, http::headervalues::contenttype::TEXT);
-        args.rpc_ctx->set_response_body(
-          jsonrpc::result_response(0, eevm::to_hex_string(result)).dump());
+      //   args.rpc_ctx->set_response_status(HTTP_STATUS_OK);
+      //   args.rpc_ctx->set_response_header(
+      //     http::headers::CONTENT_TYPE, http::headervalues::contenttype::TEXT);
+      //   args.rpc_ctx->set_response_body(
+      //     jsonrpc::result_response(0, eevm::to_hex_string(result)).dump());
 
-        // run in evm
-        auto ct = workerQueue.GetCloakTransaction(result);
-        if (ct.has_value() && ct.value()->function.complete())
-        {
-          CloakTransaction *ct_value = ct.value();
-          CLOAK_DEBUG_FMT("ct function: {}\n", ct_value->function.info());
-          ct_value->set_status(PACKAGE);
-          auto data = ct_value->function.packed_to_data();
-          MessageCall mc;
-          mc.from = mpt.from;
-          mc.to = mpt.to;
-          mc.data = to_hex_string(data);
-          CLOAK_DEBUG_FMT("ct function data: {}", mc.data);
-          auto es = make_state(args.tx);
+      //   // run in evm
+      //   auto ct = workerQueue.GetCloakTransaction(result);
+      //   if (ct.has_value() && ct.value()->function.complete())
+      //   {
+      //     CloakTransaction *ct_value = ct.value();
+      //     CLOAK_DEBUG_FMT("ct function: {}\n", ct_value->function.info());
+      //     ct_value->set_status(PACKAGE);
+      //     auto data = ct_value->function.packed_to_data();
+      //     MessageCall mc;
+      //     mc.from = mpt.from;
+      //     mc.to = mpt.to;
+      //     mc.data = to_hex_string(data);
+      //     CLOAK_DEBUG_FMT("ct function data: {}", mc.data);
+      //     auto es = make_state(args.tx);
 
-          const auto res = run_in_evm(mc, es).first;
-          CLOAK_DEBUG_FMT("run in evm, res: {}, msg: {}\n", res.output, res.exmsg);
-          if (res.er == ExitReason::threw) {
-              ct_value->set_status(FAILED);
-          } else {
-              // TODO: add succeeded status
-              // ct_value->>set_status()
-          }
-          // TODO: handle return result
-        }
+      //     const auto res = run_in_evm(mc, es).first;
+      //     CLOAK_DEBUG_FMT("run in evm, res: {}, msg: {}\n", res.output, res.exmsg);
+      //     if (res.er == ExitReason::threw) {
+      //         ct_value->set_status(FAILED);
+      //     } else {
+      //         // TODO: add succeeded status
+      //         // ct_value->>set_status()
+      //     }
+      //     // TODO: handle return result
+      //   }
 
-        return ccf::make_success("");
-      };
+      //   return ccf::make_success("");
+      // };
 
-      auto get_multiPartyStatus = [this](ccf::EndpointContext& args) {
-        const auto body_j =
-          nlohmann::json::parse(args.rpc_ctx->get_request_body());
-        auto mps = body_j.get<rpcparams::GetMultiPartyStatus>();
-        auto result = workerQueue.getMultiPartyStatus(mps.tx_hash);
+      // auto get_multiPartyStatus = [this](ccf::EndpointContext& args) {
+      //   const auto body_j =
+      //     nlohmann::json::parse(args.rpc_ctx->get_request_body());
+      //   auto mps = body_j.get<rpcparams::GetMultiPartyStatus>();
+      //   auto result = workerQueue.getMultiPartyStatus(mps.tx_hash);
 
-        args.rpc_ctx->set_response_status(HTTP_STATUS_OK);
-        args.rpc_ctx->set_response_header(
-          http::headers::CONTENT_TYPE, http::headervalues::contenttype::TEXT);
-        args.rpc_ctx->set_response_body(
-          jsonrpc::result_response(0, result).dump());
-        return;
-      };
+      //   args.rpc_ctx->set_response_status(HTTP_STATUS_OK);
+      //   args.rpc_ctx->set_response_header(
+      //     http::headers::CONTENT_TYPE, http::headervalues::contenttype::TEXT);
+      //   args.rpc_ctx->set_response_body(
+      //     jsonrpc::result_response(0, result).dump());
+      //   return;
+      // };
 
       auto get_workOrderSubmit = [](ccf::EndpointContext& args) {
         const auto body_j =
@@ -499,9 +498,9 @@ namespace evm4ccf
         ethrpc::GetTransactionReceipt::name, HTTP_GET, get_transaction_receipt)
         .install();
 
-      make_endpoint(
-        ethrpc::GetMultiPartyStatus::name, HTTP_GET, get_multiPartyStatus)
-        .install();
+      // make_endpoint(
+      //   ethrpc::GetMultiPartyStatus::name, HTTP_GET, get_multiPartyStatus)
+      //   .install();
 
       make_endpoint(
         ethrpc::WorkOrderSubmit::name, HTTP_GET, get_workOrderSubmit)
@@ -522,15 +521,15 @@ namespace evm4ccf
       make_endpoint(ethrpc::SendTransaction::name, HTTP_POST, send_transaction)
         .install();
 
-      make_endpoint(
-        ethrpc::SendPrivacyPolicy::name, HTTP_POST, send_privacy_policy)
-        .install();
+      // make_endpoint(
+      //   ethrpc::SendPrivacyPolicy::name, HTTP_POST, send_privacy_policy)
+      //   .install();
 
-      make_endpoint(
-        ethrpc::SendMultiPartyTransaction::name,
-        HTTP_POST,
-        send_multiPartyTransaction)
-        .install();
+      // make_endpoint(
+      //   ethrpc::SendMultiPartyTransaction::name,
+      //   HTTP_POST,
+      //   send_multiPartyTransaction)
+      //   .install();
 
       make_endpoint(
         "eth_getTransactionCount_Test",
@@ -552,8 +551,7 @@ namespace evm4ccf
       },
       storage("eth.storage"),
       tx_results("eth.txresults"),
-      txTables(*nwt.tables),
-      workerQueue(*nwt.tables)
+      txTables(*nwt.tables)
     // SNIPPET_END: initialization
     {
       context.get_historical_state();
