+++ src/transaction/signature_abstract.h
